Change Data type of Number in stock of movies to byte

+++ Vidly/Dtos/MovieDto.cs
@@ -18,7 +18,7 @@ namespace Vidly.Dtos
         public DateTime DateAdded { get; set; }
 
         [Range(1, 20)]
-        public short NumberInStock { get; set; }
+        public byte NumberInStock { get; set; }
 
         public byte GenreId { get; set; }
